Pulled the new createRoot logic into util function
